Tests for default implementations in traits
