+++ src/ops/tests.rs
@@ -1,0 +1,209 @@
+use super::*;
+
+use crate::test_utils;
+
+//-----------------------------------------------------------------------------
+
+// A naive integer vector for testing the default implementations in traits.
+struct NaiveVector(Vec<u64>);
+
+impl From<Vec<u64>> for NaiveVector {
+    fn from(source: Vec<u64>) -> Self {
+        NaiveVector(source)
+    }
+}
+
+impl Vector for NaiveVector {
+    type Item = u64;
+
+    fn len(&self) -> usize {
+        self.0.len()
+    }
+
+    fn width(&self) -> usize {
+        64
+    }
+
+    fn max_len(&self) -> usize {
+        usize::MAX
+    }
+}
+
+impl<'a> Access<'a> for NaiveVector {
+    type Iter = AccessIter<'a, Self>;
+
+    fn get(&self, index: usize) -> <Self as Vector>::Item {
+        self.0[index]
+    }
+
+    fn iter(&'a self) -> Self::Iter {
+        Self::Iter::new(self)
+    }
+}
+
+struct ValueIter<'a> {
+    parent: &'a NaiveVector,
+    value: u64,
+    rank: usize,
+    index: usize,
+}
+
+impl<'a> Iterator for ValueIter<'a> {
+    type Item = (usize, usize);
+
+    fn next(&mut self) -> Option<Self::Item> {
+        while self.index < self.parent.len() {
+            if self.parent.get(self.index) == self.value {
+                let result = Some((self.rank, self.index));
+                self.rank += 1; self.index += 1;
+                return result;
+            }
+            self.index += 1;
+        }
+        None
+    }
+}
+
+impl<'a> VectorIndex<'a> for NaiveVector {
+    type ValueIter = ValueIter<'a>;
+
+    fn rank(&self, index: usize, value: <Self as Vector>::Item) -> usize {
+        let index = cmp::min(index, self.len());
+        let mut result = 0;
+        for i in 0..index {
+            if self.get(i) == value {
+                result += 1;
+            }
+        }
+        result
+    }
+
+    fn value_iter(&'a self, value: <Self as Vector>::Item) -> Self::ValueIter {
+        Self::ValueIter {
+            parent: self,
+            value,
+            rank: 0,
+            index: 0,
+        }
+    }
+
+    fn value_of(iter: &Self::ValueIter) -> <Self as Vector>::Item {
+        iter.value
+    }
+
+    fn select(&self, rank: usize, value: <Self as Vector>::Item) -> Option<usize> {
+        let mut found = 0;
+        for index in 0..self.len() {
+            if self.get(index) == value {
+                if found == rank {
+                    return Some(index);
+                }
+                found += 1;
+            }
+        }
+        None
+    }
+
+    fn select_iter(&'a self, rank: usize, value: <Self as Vector>::Item) -> Self::ValueIter {
+        let index = self.select(rank, value).unwrap_or(self.len());
+        Self::ValueIter { parent: self, value, rank, index, }
+    }
+}
+
+//-----------------------------------------------------------------------------
+
+#[test]
+fn access_iter() {
+    let data = test_utils::random_vector(322, 7);
+    let naive = NaiveVector::from(data.clone());
+
+    assert!(naive.iter().eq(data.iter().cloned()), "Invalid values from iterator");
+
+    let mut naive_iter = naive.iter();
+    let mut data_iter = data.iter().cloned();
+    while let Some(value) = naive_iter.next_back() {
+        assert_eq!(Some(value), data_iter.next_back(), "Invalid values from reverse iterator");
+    }
+    assert!(data_iter.next_back().is_none(), "Did not get enough values from reverse iterator");
+}
+
+#[test]
+fn access_iter_nth() {
+    let data = test_utils::random_vector(271, 6);
+    let naive = NaiveVector::from(data);
+
+    // Forward.
+    for i in 0..naive.len() {
+        assert_eq!(naive.iter().nth(i), Some(naive.get(i)), "Invalid nth({})", i);
+    }
+
+    // Backward.
+    for i in 0..naive.len() {
+        assert_eq!(naive.iter().nth_back(i), Some(naive.get(naive.len() - 1 - i)), "Invalid nth_back({})", i);
+    }
+}
+
+#[test]
+fn has_item() {
+    let width = 8;
+    let data = test_utils::random_vector(198, width);
+    let naive = NaiveVector::from(data);
+
+    for value in 0..(1 << width) {
+        let should_have = naive.iter().any(|x| x == value);
+        assert_eq!(naive.has_item(value), should_have, "Invalid has_item({})", value);
+    }
+}
+
+#[test]
+fn inverse_select() {
+    let data = test_utils::random_vector(322, 6);
+    let naive = NaiveVector::from(data);
+
+    for i in 0..naive.len() {
+        let result = naive.inverse_select(i);
+        assert!(result.is_some(), "No result for inverse_select({})", i);
+        let result = result.unwrap();
+        assert_eq!(naive.select(result.0, result.1), Some(i), "Invalid inverse_select({})", i);
+    }
+    assert!(naive.inverse_select(naive.len()).is_none(), "Got an inverse_select() result past the end");
+}
+
+#[test]
+fn pred_succ() {
+    let width = 7;
+    let data = test_utils::random_vector(179, width);
+    let naive = NaiveVector::from(data);
+
+    for value in 0..(1 << width) {
+        // Try also querying at past-the-end position.
+        for index in 0..=naive.len() {
+            let mut expected: Option<(usize, usize)> = None;
+            let limit = cmp::min(index + 1, naive.len());
+            for i in 0..limit {
+                if naive.get(i) == value {
+                    expected = match expected {
+                        Some((rank, _)) => Some((rank + 1, i)),
+                        None => Some((0, i)),
+                    };
+                }
+            }
+            assert_eq!(naive.predecessor(index, value).next(), expected, "Invalid predecessor({}, {})", index, value);
+
+            let rank = if expected.is_some() { expected.unwrap().0 + 1 } else { 0 };
+            if expected.is_none() || expected.unwrap().1 < index {
+                expected = None;
+                for i in limit..naive.len() {
+                    if naive.get(i) == value {
+                        expected = Some((rank, i));
+                        break;
+                    }
+                }
+            }
+            assert_eq!(naive.successor(index, value).next(), expected, "Invalid successor({}, {})", index, value);
+        }
+    }
+}
+
+//-----------------------------------------------------------------------------
+
